chore: refactor push notification
