+++ components/notifications/push-notification/index.ts
@@ -1,0 +1,73 @@
+import { Platform } from "react-native";
+import Constants from 'expo-constants';
+import * as Device from 'expo-device';
+
+export type TMessagePushNotication = {
+	to: string,
+	sound: string,
+	title: string,
+	body: string,
+	data?: {
+		[key: string]: string
+	}
+
+}
+
+export async function sendPushNotification(message: TMessagePushNotication) {
+	await fetch('https://exp.host/--/api/v2/push/send', {
+		method: 'POST',
+		headers: {
+			Accept: 'application/json',
+			'Accept-encoding': 'gzip, deflate',
+			'Content-Type': 'application/json',
+		},
+		body: JSON.stringify(message),
+	});
+}
+
+
+export async function registerForPushNotificationsAsync(Notifications: typeof import("expo-notifications")) {
+	if (Platform.OS === 'android') {
+		Notifications.setNotificationChannelAsync('default', {
+			name: 'default',
+			importance: Notifications.AndroidImportance.MAX,
+			vibrationPattern: [0, 250, 250, 250],
+			lightColor: '#FF231F7C',
+		});
+	}
+
+	if (Device.isDevice) {
+		const { status: existingStatus } = await Notifications.getPermissionsAsync();
+		let finalStatus = existingStatus;
+		if (existingStatus !== 'granted') {
+			const { status } = await Notifications.requestPermissionsAsync();
+			finalStatus = status;
+		}
+		if (finalStatus !== 'granted') {
+			handleRegistrationError('Permission not granted to get push token for push notification!');
+			return;
+		}
+		const projectId =
+			Constants?.expoConfig?.extra?.eas?.projectId ?? Constants?.easConfig?.projectId;
+		if (!projectId) {
+			handleRegistrationError('Project ID not found');
+		}
+		try {
+			const pushTokenString = (
+				await Notifications.getExpoPushTokenAsync({
+					projectId,
+				})
+			).data;
+			return pushTokenString;
+		} catch (e: unknown) {
+			handleRegistrationError(`${e}`);
+		}
+	} else {
+		handleRegistrationError('Must use physical device for push notifications');
+	}
+}
+
+export function handleRegistrationError(errorMessage: string) {
+	alert(errorMessage);
+	throw new Error(errorMessage);
+}
